refactor(test): move p4 exception handling test to related test suite
